日記レビュー機能を実装し、APIからデータを取得して表示するように変更
